deprecated nupp allocators
fixed can enum indentation
fixed can::Channel::rx type error

+++ libs/include/nupp/can.hpp
@@ -56,8 +56,8 @@ namespace nu {
 								CAN_FILTER_MASK mask, CAN_FILTER_MASK_TYPE mide,
 								uint32_t mask_bits);
 
-			size_t rx(void *dest, uint32_t &id);
-			size_t rx(frame::Packet &p, uint32_t &id);
+			int32_t rx(void *dest, uint32_t &id);
+			int32_t rx(frame::Packet &p, uint32_t &id);
 			int32_t tx(const void *data, size_t num_bytes, uint16_t std_id,
 					uint32_t ext_id = 0, id_type type = STANDARD_ID);
 			int32_t tx(const frame::Packet &p);
